server/requirements.txt
amqp==5.1.1
annotated-types==0.6.0
anyascii==0.3.2
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.11.2
billiard==3.6.4.0
black==23.7.0
celery==5.2.7
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
defusedxml==0.7.1
Django==4.2.6
django-admin-interface==0.26.0
django-bulk-update==2.2.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-colorfield==0.9.0
django-cors-headers==4.0.0
django-countries==7.5.1
django-enum-choices==2.1.4
django-environ==0.10.0
django-filter==23.3
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-modelcluster==6.1
django-mptt==0.15.0
django-nested-inline==0.4.6
django-permissionedforms==0.1
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-taggit==4.0.0
django-timezone-field==5.0
django-treebeard==4.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
draftjs-exporter==2.1.7
drf-yasg==1.21.7
et-xmlfile==1.1.0
filetype==1.2.0
gunicorn==21.2.0
html5lib==1.1
idna==3.4
inflection==0.5.1
install==1.3.5
isodate==0.6.1
kombu==5.3.1
l18n==2021.3
lxml==4.9.3
mypy==1.5.0
mypy-extensions==1.0.0
openpyxl==3.1.2
packaging==23.2
pathspec==0.11.2
phonenumbers==8.13.18
Pillow==10.0.0
pillow-heif==0.13.1
platformdirs==3.11.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pydantic==2.5.1
pydantic_core==2.14.3
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
telepath==0.3.1
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
vine==5.0.0
wagtail==5.2
wcwidth==0.2.8
webencodings==0.5.1
Willow==1.6.2
zeep==4.2.1
